insert newclass not done

+++ src/entities/subject.go
@@ -1,7 +1,7 @@
 package entities
 
 type Subject struct {
-	ID         uint64 `json:"_id"`
+	ID         int `json:"_id"`
 	Name       string `json:"name"`
 	Created_at string `json:"createdAt"`
 }
